feat: add devframe Vite Bridge, move Devkit/Bruno to tools/
- Integrate devframe Vite Bridge (viteDevBridge) in vite.config.ts,
  mounting Devkit at /__devkit in the dev server for agent-native
  tool access during development (static-mount mode)
- Move src/devkit/ to tools/devkit/ and src/types/bruno-converters.d.ts
  to tools/bruno-converters.d.ts (non-core tooling belongs outside src/)
- Update tsconfig.json to include tools/ directory and fix type reference
- Update all imports (package.json scripts, test imports, PLANS.md paths)
- Update CHANGELOG.md with Bruno, Devkit, Vite Bridge entries
- Pre-commit checks: lint clean, typecheck clean, 1583 tests pass
