+++ tools/devkit/mcp.ts
